pick last instead of first value
This makes it so that we prefer latest than earlier occurrences of variable
settings.

Before the change
-----------------

    $ cat > my_script.hs
    #!/bin/bash
    echo $FOO
    echo $GOO
    ^D

    $ cat > .env
    #!/bin/sh
    FOO=first
    GOO=second
    FOO=third
    ^D

    $ dotenv -f.env ./my_script.hs
    first
    second

After the change
----------------

    $ cat > my_script.hs
    #!/bin/bash
    echo $FOO
    echo $GOO
    ^D

    $ cat > .env
    #!/bin/sh
    FOO=first
    GOO=second
    FOO=third
    ^D

    $ dotenv -f.env ./my_script.hs
    third
    second

This makes dotenv-hs consistent with both Python's dotenv and Bash/Sh's source.

The implementation may look innefficient with

    reverse . nubBy (...) . reverse

however it is as efficient as the surrounding code which is _O(n^2)_ overall as
Data.List.union is _O(n^2)_ already.

The ultimate reverse may not be needed, but I have kept it so the order in
which variables are set is unchanged.