Parse multi-word command interpolations
The previous implementation only parsed alphanumeric values within
`$()`, which meant reasonable commands like `foo.sh`, `bin/foo`, or
`foo-bar` would all be rejected.

This commit does more than relax the character set, it uses the
[shellwords][] package to parse multiple words, as a shell would, then
passes the command and any arguments to `proc`.

[shellwords]: https://hackage.haskell.org/package/shellwords

This ensures correct parsing of strings like,

    echo hi\ there
    => ["echo", "hi there"]

    cmd --"this is valid"
    => ["cmd", "--this is valid"]

    cmd --this="Is \"also\" valid"
    => ["cmd", "--this=Is \"also\" valid"]

A note about security: this implementation was originally suggested as a
safer alternative to slurping up all content between `$(` and `)` and
passing it to `shell`. I now see that it is not actually any safer, as
the following example should show:

    FOO=$(echo 'if this were unsafe to pass to shell')
    BAR=$(sh -c "echo 'then this is just as unsafe'")

If we truly don't trust this input, the whole idea of interpolated
commands is suspect. The original mechanism is problematic even with the
restricted parser, and a naive parse-to-shell would be no more or less
problematic.

I think it's reasonable to trust the input, which means any of these
approaches are acceptable. In an ideal world, we'd add a switch in
`Config` to disable this feature for the security concious.

That said, shellwords does bring (at least) two conveniences:

- It correctly handles escaped or quoted ")" characters

  For example,

      FILES=$( find \( -name '*.hs' -o -name '*.hi' \) )

  Without shellwords, we'd still want to account for this ourselves, so
  I think a totally naive parser is off the table.

- It sets up the process tree without an intermediate shell

  This almost certainly doesn't matter as these are so short-lived, but
  in rare cases it can ensure better signal handling.