Don't union in neededVars
I could be missing something, but this didn't make sense to me:

If we're in this branch, then the `ambient+read` variables have the same
keys as needed, this means `read+needed` would only add those keys in
needed that are already in the `ambient`, which also means they need to
be set. Using only `read` here is functionally equivalent.

I'd also argue that actually _using_ anything from the `.env.example`
file (beyond validating the `.env`), if that were somehow happening
despite my analysis above, would be surprising to end users anyway.