Fix unexpected error with preset ENV and .env.example (#156)
* Fix building on Stackage LTS

- Bump LTS, remove extra-deps
- Relax lower bound on mtl

* Reproduce bug

* Fix `union`/`unionBy` bug

When the ambient environment is combined with the read values, this code
used `union` which compares key _and value_. This ends up duplicating
the variable in the list that's intersected (correctly on key only
there) and the difference in length results in the missing-variable
error unnecessarily. We already have a `unionEnvs` here that does the
right thing, we just have to use it.

* Don't union in neededVars

I could be missing something, but this didn't make sense to me:

If we're in this branch, then the `ambient+read` variables have the same
keys as needed, this means `read+needed` would only add those keys in
needed that are already in the `ambient`, which also means they need to
be set. Using only `read` here is functionally equivalent.

I'd also argue that actually _using_ anything from the `.env.example`
file (beyond validating the `.env`), if that were somehow happening
despite my analysis above, would be surprising to end users anyway.

* Refactor .env.example check, improve error message

I found all of the `*By` and `cmpEnv` stuff really hard to follow and
unnecessarily complex. I think it led to the recently-fixed bug.

I find this version more straight forward. Hopefully it's objectively
less complicated too; it at least reduced two conditionals to one, which
is one quantitative improvement.

This version also made it easier to identify specifically the missing
keys for the error message. The code before would report *all* keys from
the example file(s) and leave it up to the user to sort out what was
present or not. Now we can report specific keys directly. I made some
minor quality-of-life improvements to the message too.

Before:

    Missing env vars! Please, check (this/these) var(s) (is/are) set:
    THIS_ENV, THAT_ENV, MISSING_ENV, THE_OTHER_ENV, ....

After (one `.env`, one `.env.example`):

    The following variables are present in .env.example, but not set in
    the current environment or .env: MISSING_ENV

After (many `.env`s, many `.env.example`s):

    The following variables are present in one of .env.example.1,
    .env.example.2, but not set in the current environment, or any of
    .env.1, .env.2: MISSING_ENV

This should be much less confusing to end-users.

* Model unexpected empty list case in the types